move check to the far right
