+++ ui/desktop/src/components/settings_v2/providers/callbacks/RefreshActiveProviders.tsx
