+++ ui/desktop/src/components/settings_v2/providers/interfaces/OllamaMetadata.tsx
