+++ ui/desktop/src/components/settings_v2/providers/interfaces/ProviderDetails.tsx
