+++ ui/desktop/src/components/settings_v2/providers/modals/ProviderConfiguationModal.tsx
