+++ ui/desktop/src/components/settings_v2/providers/modals/configuration_modal_subcomponents/ProviderSetupForm.tsx
