+++ ui/desktop/src/components/settings_v2/providers/subcomponents/actions/TooltipWrapper.tsx
@@ -1,11 +1,6 @@
 // TooltipWrapper.tsx
 import React from 'react';
-import {
-  Tooltip,
-  TooltipTrigger,
-  TooltipContent,
-  TooltipProvider,
-} from '../../../../../ui/Tooltip';
+import { Tooltip, TooltipTrigger, TooltipContent, TooltipProvider } from '../../../../ui/Tooltip';
 import { Portal } from '@radix-ui/react-portal';
 
 interface TooltipWrapperProps {
