+++ ui/desktop/src/components/settings_v2/providers/subcomponents/utils/StringUtils.tsx
